- Moved more functions out of handler, comm, and others into extension methods - Moved more commands out of old classes into their own command classes

+++ trunk/SmaugCS/color.cs
@@ -122,7 +122,7 @@ namespace SmaugCS
 
         public static string color_str(ATTypes attype, CharacterInstance ch)
         {
-            if (ch.IsNpc() || ch.Act.IsSet((int)PlayerFlags.Ansi))
+            if (ch.IsNpc() || ch.Act.IsSet(PlayerFlags.Ansi))
                 return string.Empty;
 
             AnsiCodes code = Realm.Library.Common.EnumerationExtensions.GetEnum<AnsiCodes>((int)attype);
